Add special history component for epd and work lead
