+++ epictrack-web/src/components/workPlan/firstNations/FirstNationList.tsx
@@ -460,7 +460,7 @@ const FirstNationList = () => {
                 showGlobalFilter: true,
               }}
               onHoveredRowChange={({ ...params }) => {
-                debugger;
+                return;
               }}
             />
           </Grid>
