+++ epictrack-web/src/constants/application-constant.ts
@@ -96,9 +96,13 @@ export enum SpecialFieldEntityEnum {
   WORK = "WORK",
 }
 
-export const SpecialFields = {
+export const SPECIAL_FIELDS = Object.freeze({
   PROJECT: {
     NAME: "name",
     PROPONENT: "proponent_id",
   },
-};
+  WORK: {
+    RESPONSIBLE_EPD: "responsible_epd_id",
+    WORK_LEAD: "work_lead_id",
+  },
+});
